Revert ingredient classes, and use String type instead
