--- core/src/cs/eng1/piazzapanic/ingredients/Bun.java
@@ -1,4 +1,0 @@
-package cs.eng1.piazzapanic.ingredients;
-
-public class Bun extends Ingredient {
-}
