--- core/src/cs/eng1/piazzapanic/ingredients/Patty.java
@@ -1,5 +1,0 @@
-package cs.eng1.piazzapanic.ingredients;
-
-public class Patty extends Ingredient {
-
-}
