+++ core/src/cs/eng1/piazzapanic/stations/CookingStation.java
@@ -2,7 +2,7 @@ package cs.eng1.piazzapanic.stations;
 
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.utils.TimeUtils;
-import cs.eng1.piazzapanic.ingredients.Ingredient;
+import cs.eng1.piazzapanic.Ingredient;
 
 
 public class CookingStation extends Station {
@@ -53,9 +53,9 @@ public class CookingStation extends Station {
     }
   }
 
-  private Boolean correctIngredient(Ingredient check) {
+  private Boolean correctIngredient(Ingredient ingredientToCheck) {
     for (Ingredient item : this.validIngredients) {
-      if (item == check) {
+      if (item.getType().equals(ingredientToCheck.getType())) {
         return true;
       }
     }
