+++ core/src/cs/eng1/piazzapanic/stations/IngredientStation.java
@@ -1,7 +1,7 @@
 package cs.eng1.piazzapanic.stations;
 
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
-import cs.eng1.piazzapanic.ingredients.Ingredient;
+import cs.eng1.piazzapanic.Ingredient;
 
 public class IngredientStation extends Station {
 
@@ -9,6 +9,10 @@ public class IngredientStation extends Station {
     public IngredientStation(TextureRegion image, Ingredient ingredient){
         super(image);
         ingredientDispensed = ingredient; //What ingredient the station will give to the player.
+    }
+
+    public Ingredient dispenseNewIngredient(){
+        return new Ingredient(ingredientDispensed.getType());
     }
 
 }
